requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
supabase>=2.18.1
python-dotenv>=1.1.1
Flask-Mail==0.9.1
gunicorn==21.2.0 